Use setTimeout to wait for dom to finish painting before scrolling
There was a strange race condition happening where sometimes the browser finished painting and the hash was scrolled to correctly, other times it took a couple milliseconds and wouldn't scroll all the way down to the hash. Adding a short 300ms timeout allowed enough time for the browser to finish all painting before scrolling - resulting in consistent page appearance when linking to headings.