OpenID Connect is an Authentication System
Suggesting on changing this sentence 
"Additionally, OrderCloud has support for OpenID connect which is an authorization standard built on top of OAuth2 which enables single-sign-on and brings our grand tally up to 6 workflows."
Proposing to change the word "authorization" to "authentication". OpenID was added to OAuth 2.0 because OAuth is a authorization protocol and everyone was using their own way to authenticate using OAuth. OpenID Connect standardized the authentication process.