Turn off devtool when running in production mode
Hopefully this fixes a build error that was occurring related to source mapping in prod. Source mapping is only really necessary when debugging locally for docs so removing them on prod is okay.

Also remove the extraneous (we think) package around `source-map-support` that didn't seem to help or hurt the issue when it was removed.